Fix compilation errors in src/webkit/ when building for 64-bit Mac OS.
BUG=148664


Review URL: https://chromiumcodereview.appspot.com/10918192

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@157125 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 309b24ec457d4f5a14c7edcab1dc207e08d2d590

+++ webkit/plugins/npapi/plugin_host.cc
@@ -882,7 +882,7 @@ NPError NPN_SetValue(NPP id, NPPVariable variable, void* value) {
       return NPERR_GENERIC_ERROR;
   #if defined(OS_MACOSX)
     case NPPVpluginDrawingModel: {
-      int model = reinterpret_cast<int>(value);
+      intptr_t model = reinterpret_cast<intptr_t>(value);
       if (model == NPDrawingModelCoreGraphics ||
           ((model == NPDrawingModelInvalidatingCoreAnimation ||
             model == NPDrawingModelCoreAnimation) &&
@@ -894,7 +894,7 @@ NPError NPN_SetValue(NPP id, NPPVariable variable, void* value) {
     }
     case NPPVpluginEventModel: {
       // Only the Cocoa event model is supported.
-      int model = reinterpret_cast<int>(value);
+      intptr_t model = reinterpret_cast<intptr_t>(value);
       if (model == NPEventModelCocoa) {
         plugin->set_event_model(static_cast<NPEventModel>(model));
         return NPERR_NO_ERROR;
